feat: implement outlet component in dashboard
- add outlet component
- move settings component to /dashboard/settings page
