+++ e2e/settings.spec.ts
@@ -7,21 +7,17 @@ test.use({
   storageState: 'e2e/auth.json',
 })
 
-test('Dashboard', async ({ page, noscript, queries: { getByRole } }) => {
-  // Go to http://localhost:3000/dashboard
-  await page.goto('/dashboard')
-
+test('Settings', async ({ page, noscript, queries: { getByRole } }) => {
+  // Go to http://localhost:3000/dashboard/settings
+  await page.goto('/dashboard/settings')
   // Query phoneNumber
   const phoneNumber = await getByRole('textbox', {
     name: /nomor whatsapp/i,
   })
-
   // Fill phoneNumber
   await phoneNumber.fill('6512345678')
-
   // Press Tab
   await phoneNumber.press('Tab')
-
   // Expect visibility only when JavaScript is enabled
   if (!noscript) {
     await expect(
@@ -32,16 +28,13 @@ test('Dashboard', async ({ page, noscript, queries: { getByRole } }) => {
         .first()
     ).toBeVisible()
   }
-
   // Fill phoneNumber
   await phoneNumber.fill('+6512345678')
-
   // Click text=Save
   await Promise.all([
     page.waitForNavigation(/*{ url: 'http://localhost:3000/dashboard' }*/),
     page.click('text=Simpan'),
   ])
-
   // Expect invisibility
   await expect(
     page
